pyproject.toml
[tool.poetry]
name = "runeascend"
version = "1.0.2"
description = "Runescape Trading Platform"
authors = ["Charles Warren <ciwarren@mtu.edu>"]
include = ["runeascend", "tests"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.3"
clickhouse-driver = "^0.2.6"
pyyaml = "^6.0.1"
discord = "^2.3.2"
twisted = "^24.3.0"
kafka-python = "^2.0.2"
structlog = "^24.2.0"
requests = "^2.32.3"
faust-streaming = "^0.11.1"

[tool.poetry.scripts]
# RuneServant
trade_seeker = "runeascend.runeservant.seeker:main"
discord_bot = "runeascend.runeservant.bot:main"

# RuneSpreader 
publisher = "runeascend.runespreader.publisher:main"
archiver = "runeascend.runespreader.archiver:main"

# RuneVault
consumer = "runeascend.runevault.consumer:main"
vault = "runeascend.runevault.vault:main"

[tool.poetry.group.dev.dependencies]
deptry = "^0.16.1"
black = "^24.4.2"
isort = "^5.13.2"
pytest-cov = "^4.1.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
testcontainers = "^4.7.2"
clickhouse-migrations = "^0.7.1"
anyio = "^4.4.0"
trio = "^0.26.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80

[tool.isort]
profile = "black"
line_length = 80

[tool.deptry]
ignore = ["DEP003"]
